db_noticias_google_sheets/requirements.txt
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
pandas==1.3.3
requests==2.26.0
bs4==0.0.1